Delete SponsorshipUtils
