+++ src/herder/test/UpgradesTests.cpp
@@ -19,8 +19,8 @@
 #include "test/TestMarket.h"
 #include "test/TestUtils.h"
 #include "test/test.h"
+#include "transactions/NewSponsorshipUtils.h"
 #include "transactions/SignatureUtils.h"
-#include "transactions/SponsorshipUtils.h"
 #include "transactions/TransactionUtils.h"
 #include "util/StatusManager.h"
 #include "util/Timer.h"
@@ -30,6 +30,7 @@
 
 using namespace stellar;
 using namespace stellar::txtest;
+using namespace stellar::SponsorshipUtils;
 
 struct LedgerUpgradeableData
 {
