+++ src/transactions/TransactionFrame.cpp
@@ -19,9 +19,9 @@
 #include "ledger/LedgerTxnEntry.h"
 #include "ledger/LedgerTxnHeader.h"
 #include "main/Application.h"
+#include "transactions/NewSponsorshipUtils.h"
 #include "transactions/SignatureChecker.h"
 #include "transactions/SignatureUtils.h"
-#include "transactions/SponsorshipUtils.h"
 #include "transactions/TransactionBridge.h"
 #include "transactions/TransactionUtils.h"
 #include "util/Algorithm.h"
@@ -40,6 +40,8 @@
 
 #include <algorithm>
 #include <numeric>
+
+using namespace stellar::SponsorshipUtils;
 
 namespace stellar
 {
@@ -626,13 +628,12 @@ TransactionFrame::removeAccountSigner(AbstractLedgerTxn& ltxOuter,
         return; // probably account was removed due to merge operation
     }
 
-    auto header = ltx.loadHeader();
     auto& signers = account.current().data.account().signers;
     auto findRes = findSignerByKey(signers.begin(), signers.end(), signerKey);
     if (findRes.second)
     {
-        removeSignerWithPossibleSponsorship(ltx, header, findRes.first,
-                                            account);
+        SignerSponsorable ss(accountID, findRes.first - signers.begin());
+        ss.erase(ltx);
         ltx.commit();
     }
 }
